+++ Src/Library/changelog.md
@@ -12,6 +12,7 @@
 - response dto initialization not working with array types #225
 - unable to instantiate validators for unit tests [#info](https://discord.com/channels/933662816458645504/1017889876521267263)
 - nested schema resolving in nswag operation processor [#info](https://discord.com/channels/933662816458645504/1018565805555863572)
+- concurrent test execution bug #224
 
 ### IMPROVEMENTS
 - swagger schema resolving